android: Update

+++ unidbg-android/src/main/java/com/github/unidbg/linux/android/ArmLD.java
@@ -252,7 +252,7 @@ public class ArmLD extends Dlfcn {
                             if (log.isDebugEnabled()) {
                                 log.debug("dlsym handle=0x" + Long.toHexString(handle) + ", symbol=" + symbol.getString(0));
                             }
-                            return dlsym(emulator, handle & 0xffffffffL, symbol.getString(0));
+                            return dlsym(emulator, (int) (handle & 0xffffffffL), symbol.getString(0));
                         }
                     }).peer;
                 case "dl_unwind_find_exidx":
